+++ backend/canisters/user/impl/src/lib.rs
@@ -130,7 +130,7 @@ impl RuntimeState {
                 sender,
                 recipients: vec![recipient],
                 notification_bytes: ByteBuf::from(serialize_then_unwrap(notification)),
-                fcm_data,
+                fcm_data: Some(fcm_data),
             }))),
         })
     }