+++ backend/canisters/user/impl/src/updates/start_video_call.rs
@@ -42,10 +42,10 @@ fn start_video_call_impl(args: Args, state: &mut RuntimeState) -> OCResult {
     if !mute_notification {
         // TODO i18n
         // TODO video call notifications could display decline and answer buttons
-        let fcm_data = FcmData::builder()
-            .with_alt_title(&args.initiator_display_name, &args.initiator_username)
-            .with_body("Video call incoming...".to_string())
-            .build();
+        let fcm_data = FcmData::for_direct_chat(sender)
+            .set_body("Video call incoming...".to_string())
+            .set_body_with_alt(&args.initiator_display_name, &args.initiator_username)
+            .set_avatar_id(args.initiator_avatar_id);
 
         let notification = UserNotificationPayload::DirectMessage(DirectMessageNotification {
             sender,