+++ backend/libraries/types/src/fcm_data.rs
@@ -1,90 +1,151 @@
+use crate::{ChannelId, Chat, ChatId, CommunityId, MessageId, UserId};
 use candid::CandidType;
 use serde::{Deserialize, Serialize};
 use std::collections::HashMap;
 
 // Values relevant for the FCM notifications
-#[derive(CandidType, Serialize, Deserialize, Clone, Default, Debug)]
+#[derive(CandidType, Serialize, Deserialize, Clone, Debug)]
 pub struct FcmData {
+    #[serde(rename = "c")]
+    pub chat_id: Chat,
     #[serde(rename = "t")]
-    pub title: String,
+    pub thread_id: Option<MessageId>,
     #[serde(rename = "b")]
-    pub body: String,
+    pub body: Option<String>,
     #[serde(rename = "i")]
     pub image: Option<String>,
-    #[serde(rename = "d")]
-    pub data: Option<HashMap<String, String>>,
+    #[serde(rename = "s")]
+    pub sender_id: Option<UserId>,
+    #[serde(rename = "n")]
+    pub sender_name: Option<String>,
+    #[serde(rename = "a")]
+    pub avatar_id: Option<u128>,
 }
 
 impl FcmData {
-    pub fn builder() -> FcmDataBuilder {
-        FcmDataBuilder::new()
-    }
-}
-
-#[derive(Default)]
-pub struct FcmDataBuilder {
-    title: String,
-    body: String,
-    image: Option<String>,
-    data: HashMap<String, String>,
-}
-
-impl FcmDataBuilder {
-    fn new() -> Self {
+    fn default(chat_id: Chat) -> Self {
         Self {
-            title: "OpenChat".to_string(),
-            body: "You have a notification...".to_string(),
-            ..Self::default()
+            chat_id,
+            thread_id: None,
+            body: None,
+            image: None,
+            sender_id: None,
+            sender_name: None,
+            avatar_id: None,
         }
     }
 
-    /// Set title when we know what it will be!
-    pub fn with_title(self, title: String) -> Self {
-        Self { title, ..self }
+    pub fn for_direct_chat(direct_chat_id: UserId) -> Self {
+        Self {
+            sender_id: Some(direct_chat_id),
+            ..Self::default(Chat::Direct(direct_chat_id.into()))
+        }
     }
 
-    /// If wanted title value is optional, also provide alternative title!
-    /// Borrows the arg values, then creates copies to reduce boilerplate.
-    pub fn with_alt_title(self, title: &Option<String>, alt_title: &str) -> Self {
+    pub fn for_group(group_chat_id: ChatId) -> Self {
+        Self::default(Chat::Group(group_chat_id))
+    }
+
+    pub fn for_channel(community_id: CommunityId, channel_id: ChannelId) -> Self {
+        Self::default(Chat::Channel(community_id, channel_id))
+    }
+
+    pub fn set_thread_id(self, thread_id: MessageId) -> Self {
         Self {
-            title: title.clone().unwrap_or(alt_title.to_string()),
+            thread_id: Some(thread_id),
             ..self
         }
     }
 
-    /// When we know what the body is, and we can take ownership of the value
-    pub fn with_body(self, body: String) -> Self {
-        Self { body, ..self }
-    }
-
-    /// When the wanted value for the body is an Option, but we have an alternative
-    /// value we can provide. Borrows the arg values, then creates copies to
-    /// reduce boilerplate.
-    pub fn with_alt_body(self, body: &Option<String>, alt_body: &str) -> Self {
+    pub fn set_body(self, body: String) -> Self {
         Self {
-            body: body.clone().unwrap_or(alt_body.to_string()),
+            body: Some(body),
             ..self
         }
     }
 
-    /// Image for the notification
-    pub fn with_optional_image(self, image: Option<String>) -> Self {
+    pub fn set_body_with_alt(self, body: &Option<String>, alt_body: &str) -> Self {
+        Self {
+            body: if body.is_some() { body.clone() } else { Some(alt_body.into()) },
+            ..self
+        }
+    }
+
+    pub fn set_optional_image(self, image: Option<String>) -> Self {
         Self { image, ..self }
     }
 
-    /// Additional notification data in key & value format!
-    pub fn with_data(self, key: String, value: String) -> Self {
-        let mut data = self.data;
-        data.insert(key, value);
-        Self { data, ..self }
+    pub fn set_sender_id(self, sender_id: UserId) -> Self {
+        Self {
+            sender_id: Some(sender_id),
+            ..self
+        }
     }
 
-    pub fn build(self) -> FcmData {
-        FcmData {
-            title: self.title,
-            body: self.body,
-            image: self.image,
-            data: if self.data.is_empty() { None } else { Some(self.data) },
+    pub fn set_sender_name(self, sender_name: String) -> Self {
+        Self {
+            sender_name: Some(sender_name),
+            ..self
         }
+    }
+
+    pub fn set_sender_name_with_alt(self, sender_name: &Option<String>, alt_sender_name: &str) -> Self {
+        Self {
+            sender_name: if sender_name.is_some() { sender_name.clone() } else { Some(alt_sender_name.into()) },
+            ..self
+        }
+    }
+
+    pub fn set_avatar_id(self, avatar_id: Option<u128>) -> Self {
+        Self { avatar_id, ..self }
+    }
+
+    pub fn as_data(&self) -> HashMap<String, String> {
+        let mut map = HashMap::new();
+
+        match self.chat_id {
+            Chat::Channel(community_id, channel_id) => {
+                map.insert("type".into(), "community".into());
+                map.insert("community_id".into(), community_id.to_string());
+                map.insert("channel_id".into(), channel_id.to_string());
+            }
+            // Sender id is already initialised with the same value, so we
+            // ignore it here (only for direct chats).
+            Chat::Direct(_) => {
+                map.insert("type".into(), "direct".into());
+            }
+            Chat::Group(chat_id) => {
+                map.insert("type".into(), "group".into());
+                map.insert("chat_id".into(), chat_id.to_string());
+            }
+        }
+
+        if let Some(thread_id) = self.thread_id {
+            map.insert("thread_id".into(), thread_id.to_string());
+        }
+
+        if let Some(body) = &self.body {
+            map.insert("body".into(), body.clone());
+        }
+
+        if let Some(image) = &self.image {
+            map.insert("image".into(), image.clone());
+        }
+
+        // Initialised by default for direct chats, while can be set for
+        // group and community chats.
+        if let Some(sender_id) = &self.sender_id {
+            map.insert("sender_id".into(), sender_id.to_string());
+        }
+
+        if let Some(sender_name) = &self.sender_name {
+            map.insert("sender_name".into(), sender_name.clone());
+        }
+
+        if let Some(avatar_id) = self.avatar_id {
+            map.insert("avatar_id".into(), avatar_id.to_string());
+        }
+
+        map
     }
 }