+++ backend/canisters/community/CHANGELOG.md
@@ -12,6 +12,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Push community events to subscribed bots ([#8259](https://github.com/open-chat-labs/open-chat/pull/8259))
 - Push full chat events to subscribed bots ([#8264](https://github.com/open-chat-labs/open-chat/pull/8264))
 - Remove almost all usages of `Candid` from Community canisters ([#8267](https://github.com/open-chat-labs/open-chat/pull/8267))
+- Updated the `FcmData` interface ([8261](https://github.com/open-chat-labs/open-chat/pull/8261))
 
 ### Fixed